Change the way we handle deletes with the listing form
